apps/media-service/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.10.4
httpx==0.28.1
Pillow==11.1.0
